fix: apply multiple changes from integrations, refactor package structure, class names and implementations
