--- src/main/kotlin/app/revanced/patches/youtube/misc/playercontroller/fingerprint/CreateVideoPlayerSeekbarFingerprint.kt
@@ -1,15 +1,0 @@
-package app.revanced.patches.youtube.misc.playercontroller.fingerprint
-
-import app.revanced.patcher.annotation.Name
-import app.revanced.patcher.annotation.Version
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.layout.sponsorblock.annotations.SponsorBlockCompatibility
-
-@Name("create-video-player-seekbar-fingerprint")
-@SponsorBlockCompatibility
-@Version("0.0.1")
-object CreateVideoPlayerSeekbarFingerprint : MethodFingerprint(
-    "V",
-    strings = listOf("timed_markers_width")
-)
