--- src/main/kotlin/app/revanced/patches/youtube/misc/playercontroller/fingerprint/VideoLengthFingerprint.kt
@@ -1,29 +1,0 @@
-package app.revanced.patches.youtube.misc.playercontroller.fingerprint
-
-import app.revanced.patcher.annotation.Name
-import app.revanced.patcher.annotation.Version
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.layout.sponsorblock.annotations.SponsorBlockCompatibility
-import org.jf.dexlib2.Opcode
-
-@Name("video-length-fingerprint")
-
-@SponsorBlockCompatibility
-@Version("0.0.1")
-object VideoLengthFingerprint : MethodFingerprint(
-    opcodes = listOf(
-        Opcode.MOVE_RESULT_WIDE,
-        Opcode.CMP_LONG,
-        Opcode.IF_LEZ,
-        Opcode.IGET_OBJECT,
-        Opcode.CHECK_CAST,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_WIDE,
-        Opcode.GOTO,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_WIDE,
-        Opcode.CONST_4,
-        Opcode.INVOKE_VIRTUAL
-    )
-)
