+++ src/main/kotlin/app/revanced/patches/youtube/misc/video/information/annotation/PlayerControllerCompatibility.kt
@@ -1,0 +1,13 @@
+package app.revanced.patches.youtube.misc.video.information.annotation
+
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
+
+@Compatibility(
+    [Package(
+        "com.google.android.youtube", arrayOf("17.41.37")
+    )]
+)
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class PlayerControllerCompatibility
