+++ src/main/kotlin/app/revanced/patches/youtube/misc/video/information/fingerprints/PlayerInitFingerprint.kt
@@ -1,0 +1,17 @@
+package app.revanced.patches.youtube.misc.video.information.fingerprints
+
+import app.revanced.patcher.annotation.Name
+import app.revanced.patcher.annotation.Version
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patches.youtube.layout.sponsorblock.annotations.SponsorBlockCompatibility
+
+@Name("player-init-fingerprint")
+
+@SponsorBlockCompatibility
+@Version("0.0.1")
+object PlayerInitFingerprint : MethodFingerprint(
+    strings = listOf(
+        "playVideo called on player response with no videoStreamingData."
+    ),
+)
