--- src/main/kotlin/app/revanced/patches/youtube/misc/videoid/annotation/VideoIdCompatibility.kt
@@ -1,13 +1,0 @@
-package app.revanced.patches.youtube.misc.videoid.annotation
-
-import app.revanced.patcher.annotation.Compatibility
-import app.revanced.patcher.annotation.Package
-
-@Compatibility(
-    [Package(
-        "com.google.android.youtube", arrayOf("17.36.37", "17.41.37", "17.42.35", "17.43.36")
-    )]
-)
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class VideoIdCompatibility
